Changed guildId to be more general
The value was previously set as "MATRIX_GUILD_ID" which mapped to the
guild that I used for bot testing. This makes things more general, which
is much less confusing for anyone else trying to use the bot in a
different guild.

... Also added a snarky message for people who try to use the bot that
don't have the conductor role assigned to them